common refactor
